DP-6660 - Push datasets to data.overheid.nl
Get all datasets te be pushed, and get all datasets alreay pushed to data.overheid.nl

Check if there was a change an the data needs to be updated.

For datasets not yet present add them to data.overheid.nl

Remove datasets that are not in the list to be pushed and remove resources
that are not anymore present in the datasets.